requirements.txt
matplotlib
numpy
pyqtgraph
pyserial
python-dateutil
Pillow
cycler
packaging
contourpy
fonttools
kiwisolver
pyparsing
pyqt5
